requirements.txt
Flask==3.0.3
PyJWT==2.8.0
pymysql==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
reportlab==4.0.0
email-validator==2.0.0.post2
itsdangerous==2.1.2
